requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
click==7.1.2
dataclasses==0.7
Django==3.1.2
django-compat==1.0.15
django-cors-headers==3.5.0
django-crontab==0.7.1
djangorestframework==3.12.1
mypy-extensions==0.4.3
pathspec==0.8.0
pkg-resources==0.0.0
psycopg2==2.8.6
pytz==2020.1
regex==2020.10.11
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
